Moved Synaptics dispatch code to a separate file; removed some inactive code for development time (MAKE SURE TO RETURN IT BACK LATER)
